api_server_flask/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.0
black==22.10.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
distlib==0.3.6
exceptiongroup==1.0.0
filelock==3.8.0
flake8==4.0.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-jsonschema==0.13.0
mccabe==0.6.1
mypy-extensions==0.4.3
nose==1.3.7
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pprintpp==0.4.0
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
pytest-black==0.3.12
pytest-clarity==1.0.1
pytest-flake8==1.1.1
pytest-flask==1.2.0
python-dotenv==0.21.0
pytz==2022.2.1
rich==12.6.0
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.41
toml==0.10.2
tomli==2.0.1
tox==3.27.0
virtualenv==20.16.6
Werkzeug==2.1.2
